requirements.txt
altair==4.1.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==4.0.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
fake-useragent==0.1.11
feedparser==6.0.8
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
multitasking==0.0.9
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.1
openpyxl==3.0.7
packaging==21.0
packgenlite==1.1.3
pandas==1.3.1
pandocfilters==1.4.3
parse==1.19.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyarrow==5.0.0
pycparser==2.20
pydeck==0.6.2
pyee==8.1.0
Pygments==2.9.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-gitlab==2.10.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.0
requests==2.26.0
requests-html==0.10.0
requests-toolbelt==0.9.1
scipy==1.7.1
seaborn==0.11.1
Send2Trash==1.7.1
setuptools-scm==6.0.1
sgmllib3k==1.0.0
six==1.16.0
smmap==4.0.0
soupsieve==2.2.1
streamlit==0.85.1
tenacity==8.0.1
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
yahoo-fin==0.8.9.1
yfinance==0.1.63
